Adjusted gold in Database and Upgrade

+++ FantasySortingGame/src/main/java/com/example/fantasysortinggame/datatypes/Upgrade.java
@@ -2,16 +2,16 @@ package com.example.fantasysortinggame.datatypes;
 
 public class Upgrade {
     String ability;
-    int cost;
+    double cost;
     boolean isBought;
 
-    public Upgrade(String ability, int cost, boolean isBought) {
+    public Upgrade(String ability, double cost, boolean isBought) {
         this.ability = ability;
         this.cost = cost;
         this.isBought = isBought;
     }
 
-    public int getCost() {
+    public double getCost() {
         return cost;
     }
